sidecar/requirements.txt
grpcio==1.62.2
grpcio-tools==1.62.2
protobuf==4.25.3
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
torch==2.3.0
pytest==8.2.1
pytest-asyncio==0.23.7
